Refactor: 채팅-라이브 통합 리팩토링
