+++ src/main/java/com/ssginc/showpinglive/config/SecurityConfig.java
@@ -41,9 +41,9 @@ public class SecurityConfig {
                 .authorizeHttpRequests(auth -> auth
                         //루트는 모두 허용
                         //.requestMatchers("/", "/login", "/logout", "/css/**", "/js/**", "/images/**", "/assets/**").permitAll()
-                        .requestMatchers("/", "/login", "/api/register", "/webrtc/webrtc", "/css/**", "/js/**", "/images/**", "/assets/**").permitAll()
+                        .requestMatchers("/", "/login", "/api/register", "/webrtc/watch", "/css/**", "/js/**", "/images/**", "/assets/**").permitAll()
                         // 정적 리소스 접근 허용, 싱글톤빈으로 static주소(resources아래 폴더) 접근 제어 가능
-                        .requestMatchers("/admin/**", "/chat/chat", "/report/report").hasRole("ADMIN") //admin, chat/chat 주소 접근은 ADMIN role만 접근 가능
+                        .requestMatchers("/admin/**", "/report/report", "/webrtc/webrtc").hasRole("ADMIN") //admin, chat/chat 주소 접근은 ADMIN role만 접근 가능
                         .requestMatchers("/user/**", "/watch/history").hasAnyRole("USER", "ADMIN") //user 주소 접근은 ADMIN/USER role중 어떤 것이든 접근 가능
                         //.anyRequest().authenticated() //모두 인증필요
                         .anyRequest().permitAll() //모두 허용하고, admin과 user로 시작하는 사이트와 /chat/chat만 로그인 필요
