+++ src/main/resources/static/css/watch/stream.css
@@ -22,8 +22,17 @@
     height: 480px;
 }
 
+/* Chat Room Container */
 .chat-container {
+    position: relative;
     width: 300px;
+    height: 480px;
+    border: 1px solid #ddd;
+    border-radius: 5px;
+    overflow: hidden;
+    font-family: Arial, sans-serif;
+    display: flex;
+    flex-direction: column;
 }
 
 .product-container {
@@ -117,3 +126,232 @@ div.dragged {
     color: white;
 }
 
+/* Announcement Section */
+.announcement {
+    background-color: #ffe6e6;
+    color: #ff6666;
+    font-weight: bold;
+    padding: 10px;
+    text-align: center;
+    font-size: 14px;
+    border-bottom: 1px solid #ddd;
+}
+
+/* Chat Messages Section */
+.chat-messages {
+    flex-grow: 1;
+    /*height: calc(100% - 140px); !* 전체 높이에서 입력 영역(예: 50px)과 공지 영역(예: 50px) 등을 뺀 높이 *!*/
+    overflow-y: auto;
+    padding: 10px;
+    /*background-color: #f9f9f9;*/
+    margin-bottom: 10px;
+    font-size: 14px;
+    padding-bottom: 40px;
+}
+
+.chat-message .message {
+    color: #555; /* 메시지 텍스트 색상 */
+}
+
+.chat-message .username {
+    font-weight: bold;
+    margin-right: 5px;
+}
+
+.chat-message.admin .username {
+    color: #ff6666; /* 관리자 메시지 색상 */
+}
+
+.chat-message.user .username {
+    color: #333; /* 일반 사용자 메시지 색상 */
+}
+
+/* 최신 채팅으로 이동 버튼 */
+.scroll-button {
+    position: absolute;
+    bottom: 55px; /* 입력 영역 높이(예: 50px) 위에 약간의 여유 (10px) */
+    left: 50%;
+    transform: translateX(-50%);
+    padding: 5px 10px;
+    font-size: 12px;
+    background-color: #333;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    z-index: 10;
+    /* display는 JavaScript에서 제어 */
+}
+
+/* Input Section */
+.input-container {
+    display: flex;
+    border-top: #f9f9f9;
+    padding: 10px;
+}
+
+.input-container input {
+    flex-grow: 1;
+    border: 1px solid #ddd;
+    border-radius: 3px;
+    padding: 5px;
+}
+
+.input-container button {
+    background-color: #007bff;
+    color: white;
+    border: none;
+    border-radius: 3px;
+    padding: 5px 10px;
+}
+
+/* 채팅 메시지 래퍼 */
+.chat-wrapper {
+    position: relative;
+    height: 440px;
+    flex: 1; /* 위/아래로 늘어나도록 설정 */
+    overflow-y: auto; /* 스크롤 가능 */
+    padding-bottom: 70px; /* input-container가 겹치지 않도록 하단 여백 */
+    border: 1px solid #ddd; /* (선택) 영역 구분을 위한 스타일 */
+}
+
+/* 입력 영역: 항상 하단 중앙 고정 */
+.input-container {
+    position: absolute; /* 부모 .chat-container 기준 */
+    bottom: 0; /* 하단에 붙임 */
+    left: 50%; /* 수평 중앙 */
+    transform: translateX(-50%); /* 수평 중앙 정렬 보정 */
+    width: 100%; /* 예시, 필요에 따라 조정 */
+    display: flex;
+    background-color: #fff;
+    /*padding: 10px;*/
+    /*box-shadow: 0 -1px 5px rgba(0,0,0,0.1);*/
+}
+
+/*!* 메시지 입력 필드와 전송 버튼 *!*/
+/*#message-input {*/
+/*    flex: 1;*/
+/*    padding: 5px;*/
+/*    border: 1px solid #ccc;*/
+/*    !*border-radius: 3px;*!*/
+/*}*/
+
+/*#send-button {*/
+/*    margin-left: 10px;*/
+/*    padding: 5px 10px;*/
+/*    border: none;*/
+/*    border-radius: 3px;*/
+/*    background-color: #007bff;*/
+/*    color: #fff;*/
+/*    cursor: pointer;*/
+/*}*/
+
+
+.report-modal {
+    display: none; /* 초기에는 숨김 */
+    width: 320px;
+    margin: 50px auto; /* 화면 중앙 정렬 */
+    background-color: #ffffff;
+    border-radius: 10px;
+    padding: 20px;
+    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.2);
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 9999; /* 채팅창 위에 표시 */
+}
+
+/* 모달 배경(반투명) */
+.modal-overlay {
+    display: none; /* 초기에는 숨김 */
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.4);
+    z-index: 9998;
+}
+
+.report-modal h2 {
+    margin-top: 0;
+    margin-bottom: 10px;
+    font-size: 18px;
+    text-align: center;
+}
+
+.description {
+    font-size: 14px;
+    margin-bottom: 20px;
+}
+
+.radio-group {
+    margin-bottom: 20px;
+}
+
+.radio-label {
+    display: block;
+    margin-bottom: 10px;
+    font-size: 14px;
+    cursor: pointer;
+}
+
+.radio-label input[type="radio"] {
+    margin-right: 8px;
+}
+
+.report-content {
+    margin-bottom: 20px;
+    padding: 10px;
+    background-color: #f9f9f9;
+    border-radius: 6px;
+    font-size: 14px;
+}
+
+.button-group {
+    display: flex;
+    justify-content: flex-end;
+    gap: 10px;
+}
+
+#cancelBtn,
+#submitBtn {
+    padding: 8px 16px;
+    font-size: 14px;
+    cursor: pointer;
+    border: none;
+    border-radius: 4px;
+}
+
+#cancelBtn {
+    background-color: #ccc;
+    color: #333;
+}
+
+#submitBtn {
+    background-color: #f05d5e;
+    color: #fff;
+}
+
+.user-name {
+    font-weight: bold;
+    cursor: pointer;
+}
+
+#char-count {
+    margin-left: 5px;
+    margin-right: 5px;
+    font-size: 12px;
+    color: #666;
+}
+
+#error-message {
+    display: none;  /* 기본 상태에서는 숨김 */
+    color: red;
+    font-size: 14px;
+    margin-top: 5px;
+    font-weight: bold;
+    position: relative;  /* 또는 상황에 맞게 absolute 등으로 설정 */
+    z-index: 10;
+}
