--- src/main/resources/templates/chat/chatRoom.html
@@ -1,132 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorate="layout.html">
-<div layout:fragment="content">
-    <link rel="stylesheet" th:href="@{/webjars/bootstrap/5.3.3/dist/css/bootstrap.min.css}">
-    <link rel="stylesheet" th:href="@{/css/webrtc/kurento.css}">
-    <link rel="stylesheet" th:href="@{/webjars/ekko-lightbox/5.2.0/dist/ekko-lightbox.css}">
-    <link rel="stylesheet" th:href="@{/webjars/demo-console/1.5.1/index.css}">
-    <link rel="stylesheet" th:href="@{/css/chat/chat.css}">
-    <link rel="shortcut icon" th:href="@{/img/kurento.png}" type="image/png"/>
-
-    <script th:src="@{/webjars/jquery/3.7.1/dist/jquery.js}"></script>
-    <script th:src="@{/webjars/bootstrap/5.3.3/dist/js/bootstrap.min.js}"></script>
-    <script th:src="@{/webjars/webrtc-adapter/7.4.0/release/adapter.js}"></script>
-    <script th:src="@{/webjars/demo-console/1.5.1/index.js}"></script>
-    <script th:src="@{https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js}"></script>
-    <script th:src="@{https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js}"></script>
-
-    <script th:inline="javascript">
-        let chatRoomNo = [[${chatRoomNo}]];
-        let memberId = [[${memberId}]];
-    </script>
-
-    <script th:src="@{/js/chat/chatRoom.js}"></script>
-
-    <div class="container">
-        <!-- WebRTC Video Section -->
-        <div class="live-container">
-            <div id="video">
-                <video id="live-video" autoplay width="300px" height="480px"
-                       poster="/img/webrtc.png"></video>
-            </div>
-        </div>
-
-        <div class="chat-container">
-            <!-- 공지 영역 -->
-            <div class="announcement">
-                공지 | Live 방송 중에만 초특가 <strong>45% 세일 !!</strong>
-            </div>
-
-            <div class="chat-messages" id="chat-messages">
-                <!-- 초기 메시지 추가 가능 -->
-            </div>
-
-            <!-- 최신 채팅으로 이동 버튼 -->
-            <button id="scroll-to-latest" class="scroll-button" style="display:none;">
-                ↓ 최신 채팅으로 이동
-            </button>
-
-            <!-- 메시지 입력 영역 -->
-            <div class="input-container">
-                <input type="text" id="message-input" placeholder="메시지 입력" maxlength="200">
-                <span id="char-count">0/200 </span>
-                <button id="send-button">➤</button>
-            </div>
-        </div>
-
-        <!-- Product Section -->
-        <div class="product-container">
-            <div id="product">
-                <h1>상품화면</h1>
-            </div>
-        </div>
-    </div>
-
-    <div class="modal-overlay" id="modalOverlay" style="display:none;"></div>
-
-    <div class="report-modal" id="reportModal">
-        <h2>채팅 신고</h2>
-        <p class="description">신고 사유를 선택해 주세요.</p>
-        <form id="reportForm">
-            <div class="radio-group">
-                <label class="radio-label">
-                    <input type="radio" name="reportReason" value="음란성, 폭력성 또는 괴롭힘" required/>
-                    음란성, 폭력성 또는 괴롭힘
-                </label>
-                <label class="radio-label">
-                    <input type="radio" name="reportReason" value="비방 및 명예훼손"/>
-                    비방 및 명예훼손
-                </label>
-                <label class="radio-label">
-                    <input type="radio" name="reportReason" value="스팸 및 도배 행위"/>
-                    스팸 및 도배 행위
-                </label>
-                <label class="radio-label">
-                    <input type="radio" name="reportReason" value="광고성 또는 부적절한 발언"/>
-                    광고성 또는 부적절한 발언
-                </label>
-            </div>
-            <div class="report-content">
-                <!-- 신고 대상 정보가 동적으로 들어갈 부분 -->
-                <p class="chatContent" id="reportTargetText">신고 대상의 채팅 내용</p>
-            </div>
-            <div class="button-group">
-                <button type="button" id="cancelBtn">취소</button>
-                <button type="submit" id="submitBtn">신고하기</button>
-            </div>
-        </form>
-    </div>
-
-    <!-- Buttons for WebRTC and Chat -->
-    <br>
-    <div class="container">
-        <div class="settings-buttons">
-            <!-- WebRTC Buttons -->
-            <div>
-                <button id="presenter" class="btn btn-success" onclick="presenter()">
-                    송출시작
-                </button>
-                <button id="viewer" class="btn btn-primary" onclick="viewer()">
-                    방송시청
-                </button>
-                <button id="stop" href="#" class="btn btn-danger" onclick="stop()">
-                    송출종료
-                </button>
-            </div>
-
-            <!-- Chat Buttons -->
-            <div>
-                <button id="chatStart" class="btn btn-success" onclick="createChatRoom()">
-                    채팅시작
-                </button>
-                <button id="chatStop" href="#" class="btn btn-danger" onclick="stopChat()">
-                    채팅종료
-                </button>
-            </div>
-        </div>
-    </div>
-</div>
-</body>
-</html>
