+++ src/main/resources/templates/layout.html
@@ -38,7 +38,7 @@
                     <!-- 드롭다운 메뉴 -->
                     <ul class="dropdown-menu">
                         <li><a th:href="@{/webrtc/webrtc}">WebRTC</a></li>
-                        <li><a th:href="@{/webrtc/watch}">Watch</a></li>
+                        <li><a th:href="@{/webrtc/watch/1}">Watch</a></li>
                         <li><a th:href="@{/stream/list}">라이브</a></li>
                     </ul>
                 </li>
