+++ src/main/resources/templates/webrtc/webrtc.html
@@ -11,6 +11,14 @@
     <script src="/webjars/ekko-lightbox/5.2.0/dist/ekko-lightbox.min.js"></script>
     <script src="/webjars/webrtc-adapter/7.4.0/release/adapter.js"></script>
 
+    <script th:src="@{https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js}"></script>
+    <script th:src="@{https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js}"></script>
+
+    <script th:inline="javascript">
+        let chatRoomNo;
+        let memberId = [[${member.memberId}]];
+    </script>
+
     <script src="/js/kurento-utils.js"></script>
     <script src="/js/stream/stream.js"></script>
 
@@ -23,8 +31,25 @@
                 </div>
             </div>
             <div class="chat-container">
-                <div id="chat">
-                    <h1>채팅화면</h1>
+                <!-- 공지 영역 -->
+                <div class="announcement">
+                    공지 | Live 방송 중에만 초특가 <strong>45% 세일 !!</strong>
+                </div>
+
+                <div class="chat-messages" id="chat-messages">
+                    <!-- 초기 메시지 추가 가능 -->
+                </div>
+
+                <!-- 최신 채팅으로 이동 버튼 -->
+                <button id="scroll-to-latest" class="scroll-button" style="display:none;">
+                    ↓ 최신 채팅으로 이동
+                </button>
+
+                <!-- 메시지 입력 영역 -->
+                <div class="input-container">
+                    <input type="text" id="message-input" placeholder="메시지 입력" maxlength="200">
+                    <span id="char-count">0/200 </span>
+                    <button id="send-button">➤</button>
                 </div>
             </div>
             <div class="product-container">
@@ -32,6 +57,39 @@
                     <h1>상품화면</h1>
                 </div>
             </div>
+        </div>
+
+        <div class="report-modal" id="reportModal">
+            <h2>채팅 신고</h2>
+            <p class="description">신고 사유를 선택해 주세요.</p>
+            <form id="reportForm">
+                <div class="radio-group">
+                    <label class="radio-label">
+                        <input type="radio" name="reportReason" value="음란성, 폭력성 또는 괴롭힘" required/>
+                        음란성, 폭력성 또는 괴롭힘
+                    </label>
+                    <label class="radio-label">
+                        <input type="radio" name="reportReason" value="비방 및 명예훼손"/>
+                        비방 및 명예훼손
+                    </label>
+                    <label class="radio-label">
+                        <input type="radio" name="reportReason" value="스팸 및 도배 행위"/>
+                        스팸 및 도배 행위
+                    </label>
+                    <label class="radio-label">
+                        <input type="radio" name="reportReason" value="광고성 또는 부적절한 발언"/>
+                        광고성 또는 부적절한 발언
+                    </label>
+                </div>
+                <div class="report-content">
+                    <!-- 신고 대상 정보가 동적으로 들어갈 부분 -->
+                    <p class="chatContent" id="reportTargetText">신고 대상의 채팅 내용</p>
+                </div>
+                <div class="button-group">
+                    <button type="button" id="cancelBtn">취소</button>
+                    <button type="submit" id="submitBtn">신고하기</button>
+                </div>
+            </form>
         </div>
     </div>
